+++ examples/resources/fmc_access_category/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_access_category.example "<access_control_policy_id>,<id>"