+++ examples/resources/fmc_access_rule/import.sh
@@ -1,0 +1,1 @@
+terraform import fmc_access_rule.example "<access_control_policy_id>,<id>"