+++ internal/provider/resource_fmc_access_rules.go
@@ -24,20 +24,20 @@ import (
 	"net/url"
 	"slices"
 	"strings"
+	"sync"
 	"time"
 
 	"github.com/CiscoDevNet/terraform-provider-fmc/internal/provider/helpers"
+	"github.com/google/uuid"
+	"github.com/hashicorp/terraform-plugin-framework-validators/resourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
-	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/stringplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
-	"github.com/hashicorp/terraform-plugin-framework/tfsdk"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-plugin-log/tflog"
 	"github.com/netascode/go-fmc"
@@ -51,26 +51,25 @@ import (
 
 // Ensure provider defined types fully satisfy framework interfaces
 var (
-	_ resource.Resource                = &AccessControlPolicyResource{}
-	_ resource.ResourceWithImportState = &AccessControlPolicyResource{}
+	_ resource.Resource = &AccessRulesResource{}
 )
 
-func NewAccessControlPolicyResource() resource.Resource {
-	return &AccessControlPolicyResource{}
+func NewAccessRulesResource() resource.Resource {
+	return &AccessRulesResource{}
 }
 
-type AccessControlPolicyResource struct {
+type AccessRulesResource struct {
 	client *fmc.Client
 }
 
-func (r *AccessControlPolicyResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
-	resp.TypeName = req.ProviderTypeName + "_access_control_policy"
+func (r *AccessRulesResource) Metadata(ctx context.Context, req resource.MetadataRequest, resp *resource.MetadataResponse) {
+	resp.TypeName = req.ProviderTypeName + "_access_rules"
 }
 
-func (r *AccessControlPolicyResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (r *AccessRulesResource) Schema(ctx context.Context, req resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		// This description is used by the documentation generator and the language server.
-		MarkdownDescription: helpers.NewAttributeDescription("This resource manages Access Control Policy (ACP) with corresponding Access Rules and Categories.").String,
+		MarkdownDescription: helpers.NewAttributeDescription("This is an early access resource and it's behaviour may change in future releases.\n This resource manages Access Rules in Access Control Policies in bulk.\n Order of the rules is meant to be preserved within the resource, however *not* between multiple `fmc_access_rules` resources that create rules within a single category/section.\n Any change to the rule set will trigger recreation of all the rules. This is done to preserve the order of the rules. This usually means, that re-created rules will be placed at the end of the policy section/category.\n").String,
 
 		Attributes: map[string]schema.Attribute{
 			"id": schema.StringAttribute{
@@ -87,115 +86,31 @@ func (r *AccessControlPolicyResource) Schema(ctx context.Context, req resource.S
 					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"name": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Name of the Access Control Policy.").String,
+			"access_control_policy_id": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Id of the Access Control Policy.").String,
 				Required:            true,
-			},
-			"type": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Type of the object; this value is always 'AccessPolicy'.").String,
-				Computed:            true,
 				PlanModifiers: []planmodifier.String{
-					stringplanmodifier.UseStateForUnknown(),
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"description": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Description of the Access Control Policy.").String,
+			"category_name": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("Name of the category that owns this rule. Either 'section' or 'category_name' can be set.").String,
 				Optional:            true,
-			},
-			"default_action": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Action to be taken, when traffic does not match any Access Rule.").AddStringEnumDescription("BLOCK", "TRUST", "PERMIT", "NETWORK_DISCOVERY", "INHERIT_FROM_PARENT").String,
-				Required:            true,
-				Validators: []validator.String{
-					stringvalidator.OneOf("BLOCK", "TRUST", "PERMIT", "NETWORK_DISCOVERY", "INHERIT_FROM_PARENT"),
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"default_action_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Id of the default action.").String,
-				Computed:            true,
-			},
-			"default_action_log_begin": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Log events at the beginning of the connection.").AddDefaultValueDescription("false").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             booldefault.StaticBool(false),
-			},
-			"default_action_log_end": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Log events at the end of the connection.").AddDefaultValueDescription("false").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             booldefault.StaticBool(false),
-			},
-			"default_action_send_events_to_fmc": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Send events to the Firepower Management Center event viewer.").AddDefaultValueDescription("false").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             booldefault.StaticBool(false),
-			},
-			"default_action_send_syslog": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Send events to a syslog server.").String,
-				Optional:            true,
-			},
-			"default_action_syslog_config_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Id of the syslog config. Can be set only when default_action_send_syslog is true and either default_action_log_begin or default_action_log_end is true. If not set, the default policy syslog configuration in Access Control Logging applies.").String,
-				Optional:            true,
-			},
-			"prefilter_policy_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Id of the Prefilter Policy.").String,
-				Optional:            true,
-			},
-			"default_action_syslog_severity": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Override the Severity of syslog alerts.").AddStringEnumDescription("ALERT", "CRIT", "DEBUG", "EMERG", "ERR", "INFO", "NOTICE", "WARNING").String,
+			"section": schema.StringAttribute{
+				MarkdownDescription: helpers.NewAttributeDescription("The section of the policy to which the rule belongs. Either 'section' or 'category_name' can be set.").AddStringEnumDescription("default", "mandatory").String,
 				Optional:            true,
 				Validators: []validator.String{
-					stringvalidator.OneOf("ALERT", "CRIT", "DEBUG", "EMERG", "ERR", "INFO", "NOTICE", "WARNING"),
+					stringvalidator.OneOf("default", "mandatory"),
+				},
+				PlanModifiers: []planmodifier.String{
+					stringplanmodifier.RequiresReplace(),
 				},
 			},
-			"default_action_snmp_config_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Id of the SNMP alert. Can be set only when either default_action_log_begin or default_action_log_end is true.").String,
-				Optional:            true,
-			},
-			"default_action_intrusion_policy_id": schema.StringAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Id of the Intrusion Policy. Cannot be set when default action is BLOCK, TRUST, NETWORK_DISCOVERY.").String,
-				Optional:            true,
-			},
-			"manage_categories": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Should this resource manage Access Policy Categories. For Data Sources this defaults to `false` (Categories are not read).").AddDefaultValueDescription("true").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             booldefault.StaticBool(true),
-			},
-			"categories": schema.ListNestedAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Ordered list of categories.").String,
-				Optional:            true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"id": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Id of the Category.").String,
-							Computed:            true,
-						},
-						"name": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Name of the Category.").String,
-							Required:            true,
-						},
-						"section": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("The section of the policy to which the category belongs. Categories must be ordered so that entire section 'mandatory' comes above the section 'default'. If you use inheritance, the mandatory section applies before child policy's own rules, while the default section applies after child policy's own rules.").AddStringEnumDescription("default", "mandatory").AddDefaultValueDescription("default").String,
-							Optional:            true,
-							Computed:            true,
-							Validators: []validator.String{
-								stringvalidator.OneOf("default", "mandatory"),
-							},
-							Default: stringdefault.StaticString("default"),
-						},
-					},
-				},
-			},
-			"manage_rules": schema.BoolAttribute{
-				MarkdownDescription: helpers.NewAttributeDescription("Should this resource manage Access Rules. For Data Sources this defaults to `false` (Access Rules are not read).").AddDefaultValueDescription("true").String,
-				Optional:            true,
-				Computed:            true,
-				Default:             booldefault.StaticBool(true),
-			},
-			"rules": schema.ListNestedAttribute{
+			"items": schema.ListNestedAttribute{
 				MarkdownDescription: helpers.NewAttributeDescription("Ordered list of Access Rules. Rules must be sorted in the order of the corresponding categories, if they have `category_name`. Uncategorized non-mandatory rules must be below all other rules.").String,
 				Optional:            true,
 				NestedObject: schema.NestedAttributeObject{
@@ -214,17 +129,6 @@ func (r *AccessControlPolicyResource) Schema(ctx context.Context, req resource.S
 						"name": schema.StringAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("Name of the Access Rule. This name needs to be uqique within the policy.").String,
 							Required:            true,
-						},
-						"category_name": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("Name of the category that owns this rule (`name` from `categories` list).").String,
-							Optional:            true,
-						},
-						"section": schema.StringAttribute{
-							MarkdownDescription: helpers.NewAttributeDescription("The section of the policy to which the rule belongs. Can only be used when the `category_name` is null. Rules must be ordered so that entire section 'mandatory' comes above the section 'default'. Null value means 'default'. If you use inheritance, the mandatory section applies before child policy's own rules, while the default section applies after child policy's own rules.").AddStringEnumDescription("default", "mandatory").String,
-							Optional:            true,
-							Validators: []validator.String{
-								stringvalidator.OneOf("default", "mandatory"),
-							},
 						},
 						"enabled": schema.BoolAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("Indicates whether the access rule is in effect (true) or not (false).").AddDefaultValueDescription("true").String,
@@ -717,7 +621,7 @@ func (r *AccessControlPolicyResource) Schema(ctx context.Context, req resource.S
 	}
 }
 
-func (r *AccessControlPolicyResource) Configure(_ context.Context, req resource.ConfigureRequest, _ *resource.ConfigureResponse) {
+func (r *AccessRulesResource) Configure(_ context.Context, req resource.ConfigureRequest, _ *resource.ConfigureResponse) {
 	if req.ProviderData == nil {
 		return
 	}
@@ -727,42 +631,25 @@ func (r *AccessControlPolicyResource) Configure(_ context.Context, req resource.
 
 // End of section. //template:end model
 
-var _ resource.ResourceWithValidateConfig = &AccessControlPolicyResource{}
+// Mutex to sync fmc_access_rules creation
+// Since creation of the access rules may be split into multible bulks,
+// we need to ensure that no other fmc_access_rules resource is trying to create access rules at the same time.
+var accessRulesCreateMu sync.Mutex
 
-func (p *AccessControlPolicyResource) ValidateConfig(ctx context.Context, req resource.ValidateConfigRequest, resp *resource.ValidateConfigResponse) {
-	var data AccessControlPolicy
-
-	diags := req.Config.Get(ctx, &data)
-	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
-		return
-	}
-
-	// If manage_rules is not set, it defaults to true.
-	if !data.ManageRules.IsNull() && !data.ManageRules.ValueBool() && len(data.Rules) > 0 {
-		resp.Diagnostics.AddAttributeError(
-			path.Root("rules"),
-			"Conflicting Configuration",
-			"Rules cannot be defined when manage_rules is set to false",
-		)
-		return
-	}
-
-	// If manage_categories is not set, it defaults to true.
-	if !data.ManageCategories.IsNull() && !data.ManageCategories.ValueBool() && len(data.Categories) > 0 {
-		resp.Diagnostics.AddAttributeError(
-			path.Root("categories"),
-			"Conflicting Configuration",
-			"Categories cannot be defined when manage_categories is set to false",
-		)
-		return
+func (r AccessRulesResource) ConfigValidators(ctx context.Context) []resource.ConfigValidator {
+	return []resource.ConfigValidator{
+		resourcevalidator.Conflicting(
+			path.MatchRoot("category_name"),
+			path.MatchRoot("section"),
+		),
 	}
 }
 
-func (r *AccessControlPolicyResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
-	var plan AccessControlPolicy
+func (r *AccessRulesResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
+	var plan AccessRules
 
 	// Read plan
-	plan, diags := NewValidAccessControlPolicy(ctx, req.Plan)
+	diags := req.Plan.Get(ctx, &plan)
 	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
 		return
 	}
@@ -773,55 +660,26 @@ func (r *AccessControlPolicyResource) Create(ctx context.Context, req resource.C
 		reqMods = append(reqMods, fmc.DomainName(plan.Domain.ValueString()))
 	}
 
+	// Create new UUID for the bulk resource
+	plan.Id = types.StringValue(uuid.NewString())
+
+	// Create rules
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Create", plan.Id.ValueString()))
-
-	planBody := plan.toBody(ctx, AccessControlPolicy{})
-
-	// Create object
-	body := planBody
-	body, _ = sjson.Delete(body, "dummy_manage_categories")
-	body, _ = sjson.Delete(body, "dummy_categories")
-	body, _ = sjson.Delete(body, "dummy_manage_rules")
-	body, _ = sjson.Delete(body, "dummy_rules")
-
-	res, err := r.client.Post(plan.getPath(), body, reqMods...)
+	err := r.createRulesAt(ctx, plan, &plan, reqMods...)
 	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST/PUT), got error: %s, %s", err, res.String()))
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST/PUT), got error: %v", err))
 		return
 	}
-	plan.Id = types.StringValue(res.Get("id").String())
+	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished successfully", plan.Id.ValueString()))
 
-	read, err := r.client.Get(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString()), reqMods...)
-	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, read.String()))
-
-		res, err := r.client.Delete(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString()), reqMods...)
-		if err != nil {
-			resp.Diagnostics.AddWarning("Client Error", fmt.Sprintf("Also, cannot DELETE a hanging policy object, got error: %s, %s", err, res.String()))
-		}
-		return
-	}
-
-	plan.fromBodyUnknowns(ctx, read)
-
-	state := plan
-	state.Rules = nil
-	state.Categories = nil
-
-	state, diags = r.updateSubresources(ctx, req.Plan, plan, planBody, tfsdk.State{}, state)
+	diags = resp.State.Set(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
-
-	// On error we do Set anyway. Terraform taints our resource, and the next run is responsible to call Delete() for us.
-	diags = resp.State.Set(ctx, &state)
-	resp.Diagnostics.Append(diags...)
-
-	tflog.Debug(ctx, fmt.Sprintf("%s: Create finished", state.Id.ValueString()))
 
 	helpers.SetFlagImporting(ctx, false, resp.Private, &resp.Diagnostics)
 }
 
-func (r *AccessControlPolicyResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
-	var state AccessControlPolicy
+func (r *AccessRulesResource) Read(ctx context.Context, req resource.ReadRequest, resp *resource.ReadResponse) {
+	var state AccessRules
 
 	// Read state
 	diags := req.State.Get(ctx, &state)
@@ -835,57 +693,106 @@ func (r *AccessControlPolicyResource) Read(ctx context.Context, req resource.Rea
 		reqMods = append(reqMods, fmc.DomainName(state.Domain.ValueString()))
 	}
 
+	// Create URL path for the request
+	var ruleNames strings.Builder
+	var bulks []string
+	var resAccessRules string = ""
+	var counter int = 0
+	for i := 0; i < len(state.Items); i++ {
+		counter += 1
+		if ruleNames.Len() > 0 {
+			ruleNames.WriteString(",")
+		}
+		ruleNames.WriteString(state.Items[i].Name.ValueString())
+		// Values below are estimated based on experience with FMC API.
+		if ruleNames.Len() >= 2500 || counter%250 == 0 {
+			bulks = append(bulks, ruleNames.String())
+			ruleNames.Reset()
+			counter = 0
+		}
+	}
+	if ruleNames.Len() > 0 {
+		bulks = append(bulks, ruleNames.String())
+	}
+
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Read", state.Id.ValueString()))
+	for _, bulk := range bulks {
+		urlPath := state.getPath() + "?expanded=true&limit=1000&filter=name:" + url.QueryEscape(bulk)
 
-	resGet, err := r.client.Get(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString()), reqMods...)
-	if err != nil && strings.Contains(err.Error(), "StatusCode 404") {
-		resp.State.RemoveResource(ctx)
-		return
-	} else if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, resGet.String()))
-		return
-	}
-
-	// Prepare json string to be filled in with categories and rules, that come from separate endpoints.
-	s := resGet.String()
-
-	// Get categories, if we manage them
-	if !state.ManageCategories.IsUnknown() && state.ManageCategories.ValueBool() {
-		resCats, err := r.client.Get(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString())+"/categories?expanded=true", reqMods...)
+		// Read Access Rules
+		resTemp, err := r.client.Get(urlPath, reqMods...)
 		if err != nil {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, resGet.String()))
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, resAccessRules))
 			return
 		}
-		replaceCats := resCats.Get("items").String()
-		if replaceCats == "" {
-			replaceCats = "[]"
+		if resAccessRules == "" {
+			resAccessRules = resTemp.String()
+		} else {
+			items := gjson.Get(resTemp.String(), "items")
+			if !items.Exists() {
+				continue
+			}
+			if resItems := items.String()[1 : len(items.String())-1]; resItems != "" {
+				resAccessRules, _ = sjson.SetRaw(resAccessRules, "items.-1", resItems)
+			}
 		}
-		s, _ = sjson.SetRaw(s, "dummy_categories", replaceCats)
 	}
 
-	// Get rules, if we manage them
-	if !state.ManageRules.IsUnknown() && state.ManageRules.ValueBool() {
-		resRules, err := r.client.Get(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString())+"/accessrules?expanded=true", reqMods...)
-		if err != nil {
-			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve object (GET), got error: %s, %s", err, resGet.String()))
-			return
+	var tmp string
+
+	gjson.AddModifier("case", func(json, arg string) string {
+		if arg == "lower" {
+			return strings.ToLower(json)
 		}
-		replaceRules := resRules.Get("items").String()
-		if replaceRules == "" {
-			replaceRules = "[]"
+		return json
+	})
+
+	if !state.CategoryName.IsUnknown() && state.CategoryName.ValueString() != "" {
+		var category string
+		// Get the categories from the response
+		res := gjson.Get(resAccessRules, "items.#.metadata.category")
+		// Check if all categories are the same
+		unique := make(map[string]struct{})
+		for _, item := range res.Array() {
+			unique[item.String()] = struct{}{}
 		}
-		s, _ = sjson.SetRaw(s, "dummy_rules", replaceRules)
+		if len(unique) != 1 {
+			// If there are no or multiple categories, set a warning that would trigger resource recreation
+			category = "[WARN] Inconsistent categories detected"
+		} else {
+			category = res.Array()[0].String()
+		}
+		// Assign the category to the temporary JSON string
+		tmp, _ = sjson.Set(tmp, "category", category)
+	} else if !state.Section.IsUnknown() && state.Section.ValueString() != "" {
+		var section string
+		// Get all sections from the response
+		res := gjson.Get(resAccessRules, "items.#.metadata.section|@case:lower")
+		// Check if all sections are the same
+		unique := make(map[string]struct{})
+		for _, item := range res.Array() {
+			unique[item.String()] = struct{}{}
+		}
+		if len(unique) != 1 {
+			// If there are no or multiple sections, set a warning that would trigger resource recreation
+			section = "[WARN] Inconsistent sections detected"
+		} else {
+			section = res.Array()[0].String()
+		}
+		// Assign the section to the temporary JSON string
+		tmp, _ = sjson.Set(tmp, "section", section)
 	}
 
-	res := gjson.Parse(s)
+	// Merge computed category/section into the response
+	tmp, _ = sjson.SetRaw(tmp, "items", gjson.Get(resAccessRules, "items").String())
+	res := gjson.Parse(tmp)
+
+	tflog.Debug(ctx, fmt.Sprintf("res: %s", res.String()))
 
 	imp, diags := helpers.IsFlagImporting(ctx, req)
 	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
 		return
 	}
-
-	manageCategories := state.ManageCategories
-	manageRules := state.ManageRules
 
 	// After `terraform import` we switch to a full read.
 	if imp {
@@ -893,10 +800,6 @@ func (r *AccessControlPolicyResource) Read(ctx context.Context, req resource.Rea
 	} else {
 		state.fromBodyPartial(ctx, res)
 	}
-	state.adjustFromBody(ctx, res)
-
-	state.ManageCategories = manageCategories
-	state.ManageRules = manageRules
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Id.ValueString()))
 
@@ -906,14 +809,15 @@ func (r *AccessControlPolicyResource) Read(ctx context.Context, req resource.Rea
 	helpers.SetFlagImporting(ctx, false, resp.Private, &resp.Diagnostics)
 }
 
-func (r *AccessControlPolicyResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
-	var state AccessControlPolicy
+func (r *AccessRulesResource) Update(ctx context.Context, req resource.UpdateRequest, resp *resource.UpdateResponse) {
+	var plan, state AccessRules
 
 	// Read plan
-	plan, diags := NewValidAccessControlPolicy(ctx, req.Plan)
+	diags := req.Plan.Get(ctx, &plan)
 	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
 		return
 	}
+
 	// Read state
 	diags = req.State.Get(ctx, &state)
 	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
@@ -928,124 +832,113 @@ func (r *AccessControlPolicyResource) Update(ctx context.Context, req resource.U
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Update", plan.Id.ValueString()))
 
-	planBody := plan.toBody(ctx, state)
-	body := planBody
-	body, _ = sjson.Delete(body, "dummy_manage_categories")
-	body, _ = sjson.Delete(body, "dummy_categories")
-	body, _ = sjson.Delete(body, "dummy_manage_rules")
-	body, _ = sjson.Delete(body, "dummy_rules")
-
-	res, err := r.client.Put(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString()), body, reqMods...)
+	err := r.truncateRulesAt(ctx, &state, 0, reqMods...)
 	if err != nil {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (PUT), got error: %s, %s", err, res.String()))
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object, got error: %v", err))
 		return
 	}
 
-	plan.fromBodyUnknowns(ctx, res)
+	err = r.createRulesAt(ctx, plan, &plan, reqMods...)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to configure object (POST/PUT), got error: %v", err))
+		return
+	}
 
-	// Most of attribs are set as planned, except Rules and Categories which we'll do below.
-	orig := state
-	state = plan
-	state.Rules, state.Categories = orig.Rules, orig.Categories
+	tflog.Debug(ctx, fmt.Sprintf("%s: Update finished successfully", plan.Id.ValueString()))
 
-	state, diags = r.updateSubresources(ctx, req.Plan, plan, planBody, req.State, state)
-	resp.Diagnostics.Append(diags...)
-
-	tflog.Debug(ctx, fmt.Sprintf("%s: Update finished", plan.Id.ValueString()))
-
-	diags = resp.State.Set(ctx, &state)
+	diags = resp.State.Set(ctx, &plan)
 	resp.Diagnostics.Append(diags...)
 }
 
-// updateSubresources returns a coherent state whether it fails or succeeds. Caller should always set that state
-// into the Response (UpdateResponse, CreateResponse, ...), otherwise the API's UUIDs may go out-of-sync with
-// terraform.tfstate, which is always a big user-facing problem.
-func (r *AccessControlPolicyResource) updateSubresources(ctx context.Context, tfsdkPlan tfsdk.Plan, plan AccessControlPolicy, planBody string, tfsdkState tfsdk.State, state AccessControlPolicy) (AccessControlPolicy, diag.Diagnostics) {
-	var diags diag.Diagnostics
+func (r *AccessRulesResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
+	var state AccessRules
 
-	p := gjson.Parse(planBody)
-	bodyCats := p.Get("dummy_categories").Array()
-	bodyRules := p.Get("dummy_rules")
+	// Read state
+	diags := req.State.Get(ctx, &state)
+	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
+		return
+	}
 
 	// Set request domain if provided
 	reqMods := [](func(*fmc.Req)){}
-	if !plan.Domain.IsNull() && plan.Domain.ValueString() != "" {
-		reqMods = append(reqMods, fmc.DomainName(plan.Domain.ValueString()))
+	if !state.Domain.IsNull() && state.Domain.ValueString() != "" {
+		reqMods = append(reqMods, fmc.DomainName(state.Domain.ValueString()))
 	}
 
-	keptCats, keptRules := r.countKept(ctx, state, plan)
-
-	// Remove rules, if we manage them
-	if !plan.ManageRules.IsUnknown() && plan.ManageRules.ValueBool() {
-		err := r.truncateRulesAt(ctx, &state, keptRules, reqMods...)
-		if err != nil {
-			diags.AddError("Client Error", err.Error())
-			return state, diags
-		}
+	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Delete", state.Id.ValueString()))
+	err := r.truncateRulesAt(ctx, &state, 0, reqMods...)
+	if err != nil {
+		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object, got error: %v", err))
+		return
 	}
 
-	// Remove categories, if we manage them
-	if !plan.ManageCategories.IsUnknown() && plan.ManageCategories.ValueBool() {
-		err := r.truncateCatsAt(ctx, &state, keptCats, reqMods...)
-		if err != nil {
-			diags.AddError("Client Error", err.Error())
-			return state, diags
-		}
-	}
+	tflog.Debug(ctx, fmt.Sprintf("%s: Delete finished successfully", state.Id.ValueString()))
 
-	if len(plan.Categories) == 0 {
-		state.Categories = plan.Categories
-	}
-
-	if len(plan.Rules) == 0 {
-		state.Rules = plan.Rules
-	}
-
-	// Recreate categories, if we manage them
-	if !plan.ManageCategories.IsUnknown() && plan.ManageCategories.ValueBool() {
-		err := r.createCatsAt(ctx, plan, bodyCats, keptCats, &state, reqMods...)
-		if err != nil {
-			diags.AddError("Client Error", err.Error())
-			return state, diags
-		}
-	}
-
-	// Recreate rules, if we manage the
-	if !plan.ManageCategories.IsNull() && plan.ManageRules.ValueBool() {
-		err := r.createRulesAt(ctx, plan, bodyRules.Array(), keptRules, &state, reqMods...)
-		if err != nil {
-			diags.AddError("Client Error", err.Error())
-			return state, diags
-		}
-	}
-
-	return state, diags
+	resp.State.RemoveResource(ctx)
 }
 
-// countKept compares the state with the plan starting from index 0, and returns:
-//
-// how many categories to keep: they remain identical as to content and order
-//
-// how many rules to keep:
-//
-// - kept rules must belong to some category that is itself kept,
-//
-// - and must themselves remain identical as to id, content, and order
-func (r *AccessControlPolicyResource) countKept(ctx context.Context, state, plan AccessControlPolicy) (int, int) {
-	return 0, 0 // TODO
+// Section below is generated&owned by "gen/generator.go". //template:begin import
+// End of section. //template:end import
+
+func (r *AccessRulesResource) createRulesAt(ctx context.Context, plan AccessRules, state *AccessRules, reqMods ...func(*fmc.Req)) error {
+	var idx = 0
+	bulk := plan
+	bulk.Items = make([]AccessRulesItems, 0, bulkSizeCreate)
+	state.Items = state.Items[:0] // reset state items
+
+	// Mutex ensures that all rules, even if split into multiple bulks, are not mixed with
+	// other rules being created at the same time.
+	accessRulesCreateMu.Lock()
+	defer accessRulesCreateMu.Unlock()
+
+	// iterate over all items
+	for _, v := range plan.Items {
+		// count loops
+		idx++
+
+		// add object to current bulk
+		bulk.Items = append(bulk.Items, v)
+
+		// If bulk size was reached or all entries have been processed
+		if idx%bulkSizeCreate == 0 || idx == len(plan.Items) {
+
+			// Create object
+			body := bulk.toBody(ctx, AccessRules{})
+			body = gjson.Get(body, "items").String()
+
+			// Execute request
+			urlParams := "?bulk=true"
+			if c := plan.CategoryName.ValueString(); c != "" {
+				urlParams += "&category=" + url.QueryEscape(c)
+			} else if s := plan.Section.ValueString(); s != "" {
+				urlParams += "&section=" + url.QueryEscape(s)
+			}
+
+			res, err := r.client.Post(plan.getPath()+urlParams, body, reqMods...)
+			if err != nil {
+				return err
+			}
+
+			// Read result and save it to state
+			bulk.fromBodyUnknowns(ctx, res)
+			state.Items = append(state.Items, bulk.Items...)
+			bulk.Items = bulk.Items[:0] // reset bulk items
+		}
+	}
+	return nil
 }
 
-func (r *AccessControlPolicyResource) truncateRulesAt(ctx context.Context, state *AccessControlPolicy, kept int, reqMods ...func(*fmc.Req)) error {
+func (r *AccessRulesResource) truncateRulesAt(ctx context.Context, state *AccessRules, kept int, reqMods ...func(*fmc.Req)) error {
 	var b strings.Builder
 	var bulks []string
 	var counts []int
 	count := 0
 
-	for i := kept; i < len(state.Rules); i++ {
+	for i := kept; i < len(state.Items); i++ {
 		if b.Len() != 0 {
 			b.WriteString(",")
 		}
-		b.WriteString(state.Rules[i].Id.ValueString())
+		b.WriteString(state.Items[i].Id.ValueString())
 		count++
 		if b.Len() >= maxUrlParamLength {
 			bulks = append(bulks, b.String())
@@ -1065,169 +958,13 @@ func (r *AccessControlPolicyResource) truncateRulesAt(ctx context.Context, state
 	}()
 
 	for i, bulk := range bulks {
-		res, err := r.client.Delete(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString())+
-			"/accessrules?bulk=true&filter=ids:"+url.QueryEscape(bulk), reqMods...)
+		res, err := r.client.Delete(state.getPath()+"?bulk=true&filter=ids:"+url.QueryEscape(bulk), reqMods...)
 		if err != nil {
-			return fmt.Errorf("Failed to bulk-delete rules, got error: %v, %s", err, res.String())
+			return fmt.Errorf("failed to bulk-delete rules, got error: %v, %s", err, res.String())
 		}
 
-		state.Rules = slices.Delete(state.Rules, kept, kept+counts[i])
+		state.Items = slices.Delete(state.Items, kept, kept+counts[i])
 	}
 
 	return nil
 }
-
-func (r *AccessControlPolicyResource) truncateCatsAt(ctx context.Context, state *AccessControlPolicy, kept int, reqMods ...func(*fmc.Req)) error {
-	for i := len(state.Categories) - 1; i >= kept; i-- {
-		res, err := r.client.Delete(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString())+
-			"/categories/"+url.QueryEscape(state.Categories[i].Id.ValueString()), reqMods...)
-		if err != nil {
-			return fmt.Errorf("Failed to delete a category, got error: %v, %s", err, res)
-		}
-
-		state.Categories[i] = AccessControlPolicyCategories{}
-		state.Categories = state.Categories[:i]
-	}
-	return nil
-}
-
-func (r *AccessControlPolicyResource) createCatsAt(ctx context.Context, plan AccessControlPolicy, body []gjson.Result, startIndex int, state *AccessControlPolicy, reqMods ...func(*fmc.Req)) error {
-	for i := startIndex; i < len(plan.Categories); i++ {
-		cat := body[i].String()
-		cat, _ = sjson.Delete(cat, "id")
-		cat, _ = sjson.Delete(cat, "metadata.section")
-		params := ""
-		if s := plan.Categories[i].Section.ValueString(); s != "" {
-			params = "?section=" + url.QueryEscape(s)
-		}
-		res, err := r.client.Post(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString())+
-			"/categories"+params, cat, reqMods...)
-		if err != nil {
-			return fmt.Errorf("Failed to create a category (POST), got error: %v, %s", err, res)
-		}
-
-		item := plan.Categories[i]
-		item.Id = types.StringValue(res.Get("id").String())
-
-		if len(state.Categories) <= i {
-			state.Categories = append(state.Categories, item)
-		} else {
-			state.Categories[i] = item
-		}
-	}
-
-	return nil
-}
-
-// createRulesAt creates rules plan.Rules[startIndex:] (from startIndex to the end).
-// The `body` should be a marshalled `plan.Rules`.
-// Whether it succeeds fully or partially, it takes whatever has been really created and saves in the `state`.
-// The `state` and `&plan` might be either the same value or different.
-func (r *AccessControlPolicyResource) createRulesAt(ctx context.Context, plan AccessControlPolicy, body []gjson.Result, startIndex int, state *AccessControlPolicy, reqMods ...func(*fmc.Req)) error {
-	for i := startIndex; i < len(body); i++ {
-		bulk := `{"dummy_rules":[]}`
-		j := i
-		bulkCount := 0
-		bodyLength := 0
-		head := plan.Rules[i]
-		for ; i < len(body); i++ {
-			if !head.CategoryName.Equal(plan.Rules[i].CategoryName) || head.GetSection() != plan.Rules[i].GetSection() {
-				i--
-				break
-			}
-			rule := body[i].String()
-			rule, _ = sjson.Delete(rule, "id")
-			rule, _ = sjson.Delete(rule, "metadata.category")
-			rule, _ = sjson.Delete(rule, "metadata.section")
-
-			// Check if the body is too big for a single POST
-			bodyLength += len(rule)
-			if bodyLength >= maxPayloadSize {
-				i--
-				break
-			}
-
-			bulk, _ = sjson.SetRaw(bulk, "dummy_rules.-1", rule)
-
-			// Count the number of rules in the bulk
-			bulkCount++
-			if bulkCount >= bulkSizeCreate {
-				break
-			}
-		}
-
-		param := "?bulk=true"
-		if cat := head.CategoryName.ValueString(); cat != "" {
-			param += "&category=" + url.QueryEscape(cat)
-		} else if s := head.GetSection(); s != "default" {
-			param += "&section=" + url.QueryEscape(s)
-		}
-		res, err := r.client.Post(plan.getPath()+"/"+url.QueryEscape(plan.Id.ValueString())+"/accessrules"+param,
-			gjson.Parse(bulk).Get("dummy_rules").String(),
-			reqMods...)
-		if err != nil {
-			return fmt.Errorf("Failed to configure object (POST), got error: %s, %s", err, res.String())
-		}
-
-		for _, v := range res.Get("items").Array() {
-			if v.Get("name").String() == "" {
-				tflog.Debug(ctx, fmt.Sprintf("ignoring a bogus item in POST reply, it can happen: %s", v))
-				continue
-			}
-
-			item := plan.Rules[j]
-			item.Id = types.StringValue(v.Get("id").String())
-
-			if len(state.Rules) <= j {
-				state.Rules = append(state.Rules, item)
-			} else {
-				state.Rules[j] = item
-			}
-
-			j++
-		}
-	}
-
-	return nil
-}
-
-// Section below is generated&owned by "gen/generator.go". //template:begin delete
-
-func (r *AccessControlPolicyResource) Delete(ctx context.Context, req resource.DeleteRequest, resp *resource.DeleteResponse) {
-	var state AccessControlPolicy
-
-	// Read state
-	diags := req.State.Get(ctx, &state)
-	if resp.Diagnostics.Append(diags...); resp.Diagnostics.HasError() {
-		return
-	}
-
-	// Set request domain if provided
-	reqMods := [](func(*fmc.Req)){}
-	if !state.Domain.IsNull() && state.Domain.ValueString() != "" {
-		reqMods = append(reqMods, fmc.DomainName(state.Domain.ValueString()))
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("%s: Beginning Delete", state.Id.ValueString()))
-	res, err := r.client.Delete(state.getPath()+"/"+url.QueryEscape(state.Id.ValueString()), reqMods...)
-	if err != nil && !strings.Contains(err.Error(), "StatusCode 404") {
-		resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to delete object (DELETE), got error: %s, %s", err, res.String()))
-		return
-	}
-
-	tflog.Debug(ctx, fmt.Sprintf("%s: Delete finished successfully", state.Id.ValueString()))
-
-	resp.State.RemoveResource(ctx)
-}
-
-// End of section. //template:end delete
-
-// Section below is generated&owned by "gen/generator.go". //template:begin import
-
-func (r *AccessControlPolicyResource) ImportState(ctx context.Context, req resource.ImportStateRequest, resp *resource.ImportStateResponse) {
-	resource.ImportStatePassthroughID(ctx, path.Root("id"), req, resp)
-
-	helpers.SetFlagImporting(ctx, true, resp.Private, &resp.Diagnostics)
-}
-
-// End of section. //template:end import